backend/requirements.txt
alembic~=1.5.2
aiofiles==0.6.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
ecdsa==0.14.1
fastapi-utils==0.2.1
h11==0.12.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.2
passlib[bcrypt]==1.7.4
plotly==4.14.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pydantic-sqlalchemy==0.0.8.post1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose[cryptography]==3.2.0
python-multipart==0.0.5
pytz==2020.5
retrying==1.3.3
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
uvicorn==0.13.3
git+git://github.com/christinoleo/fastapi_crud_orm_connector@master#egg=fastapi-crud-orm-connector
